🌱 Complete feedback/stellar component splits (last mile)
Extracts remaining state/logic from the four target files into
custom hooks and a summary sub-component, following the existing
.parts.tsx/.utils.ts split convention introduced by #21758/#21759:

- FeedbackModal.tsx (555 -> 252 lines, 20 -> 3 hook calls): moved
  draft persistence, screenshot handling, and submission logic into
  useFeedbackDraft.ts.
- SubmitTab.tsx (594 -> 348 lines, 13 -> 2 hook calls): moved
  parent-issue linking, fullscreen preview handling, paste-to-attach,
  and the submit handler into useSubmitFormHandler.ts.
- BatchMonitorModal.tsx (360 -> 285 lines): extracted the summary
  header (progress bar + breakdown chips) into BatchSummary.tsx.
- EventModal.tsx (370 -> 203 lines, 13 -> 1 hook call): moved all
  derived data (related events/solves/activity, timeline, dirty
  state) and action handlers into useEventModalData.ts.

All four files are now below the 350-line / 10-hook-call targets
from #21781. Pure refactor, no behavior change; build and lint pass
with only pre-existing warnings.

Fixes #21781, Fixes #21800, Fixes #21801, Fixes #21802
